fix a bug in V1742 lib that convert float into unsigned int.

+++ V1742_lib.h
@@ -18,7 +18,7 @@ V1742_Event_t(const CAEN_DGTZ_EventInfo_t& ei, const CAEN_DGTZ_X742_EVENT_t& e):
 int init_V1742(int handle);
 int read_V1742(int handle, unsigned int nevents, std::vector<V1742_Event_t>& events);
 //int read_V1742(int handle);
-int writeEventToOutputBuffer_V1742(std::vector<unsigned int> *eventBuffer, CAEN_DGTZ_EventInfo_t *EventInfo, CAEN_DGTZ_X742_EVENT_t *Event);
+int writeEventToOutputBuffer_V1742(std::vector<float> *eventBuffer, CAEN_DGTZ_EventInfo_t *EventInfo, CAEN_DGTZ_X742_EVENT_t *Event);
 int stop_V1742(int handle);
 
 #endif
